Extract auth/gist UI wiring out of view.ts into gist-ui.ts
setupAuthAndGistUi grew to ~160 lines in view.ts. Move it to its own
gist-ui.ts module so view.ts can go back to being the high-level
orchestrator. The new module takes the CodeEditor instance as a
parameter, which lets us drop the cmView.view DOM hack and call
codeCM.getCode() directly. No behavior change — Cypress green.
