+++ services/frontend/src/ts/view.ts
@@ -1,11 +1,10 @@
-import { ColorMode, ExecutionResult, GistMetadata, ThemeColor, User } from './types'
+import { ColorMode, ExecutionResult, ThemeColor } from './types'
 import { combineLatest, fromEvent, Observable, of, startWith } from 'rxjs'
 import { concatMap, delay, map, tap, throttleTime } from 'rxjs/operators'
 import { GroovyConsole } from './groovy-console'
 import { compressToBase64 } from './compression'
 import { CodeEditor, OutputEditor } from './codemirror'
-import { currentUser$, loadedGist$, refreshMe, signIn, signOut } from './auth'
-import { createGist, updateGist } from './github'
+import { setupGistUi } from './gist-ui'
 
 const groovyConsole = new GroovyConsole()
 const htmlRoot = document.getElementsByTagName('html')[0]
@@ -216,170 +215,6 @@ function setupNavbarBurgerClickHandlers () {
   })
 }
 
-function setupAuthAndGistUi () {
-  const signInItem = document.getElementById('signInItem')
-  const accountItem = document.getElementById('accountItem')
-  const accountAvatar = document.getElementById('accountAvatar') as HTMLImageElement
-  const accountLogin = document.getElementById('accountLogin')
-  const signOutItem = document.getElementById('signOutItem')
-  const saveAsPublicGistControl = document.getElementById('saveAsPublicGistControl')
-  const saveAsSecretGistControl = document.getElementById('saveAsSecretGistControl')
-  const updateGistControl = document.getElementById('updateGistControl')
-  const saveAsNewGistControl = document.getElementById('saveAsNewGistControl')
-  const saveAsPublicGistBtn = document.getElementById('saveAsPublicGist')
-  const saveAsSecretGistBtn = document.getElementById('saveAsSecretGist')
-  const updateGistBtn = document.getElementById('updateGist')
-  const saveAsNewGistBtn = document.getElementById('saveAsNewGist')
-
-  const saveGistModal = document.getElementById('saveGistModal')
-  const saveGistModalTitle = document.getElementById('saveGistModalTitle')
-  const saveGistModalClose = document.getElementById('saveGistModalClose')
-  const saveGistName = document.getElementById('saveGistName') as HTMLInputElement
-  const saveGistNameError = document.getElementById('saveGistNameError')
-  const saveGistIncludeOutputField = document.getElementById('saveGistIncludeOutputField')
-  const saveGistIncludeOutput = document.getElementById('saveGistIncludeOutput') as HTMLInputElement
-  const saveGistError = document.getElementById('saveGistError')
-  const saveGistConfirm = document.getElementById('saveGistConfirm')
-  const saveGistCancel = document.getElementById('saveGistCancel')
-
-  const sessionNotification = document.getElementById('sessionNotification')
-  const sessionNotificationText = document.getElementById('sessionNotificationText')
-  const sessionNotificationClose = document.getElementById('sessionNotificationClose')
-
-  let pendingVisibility: boolean | null = null
-
-  function notifySessionExpired () {
-    sessionNotificationText.textContent = 'Your GitHub session expired. Please sign in again.'
-    sessionNotification.classList.remove('is-hidden')
-  }
-
-  sessionNotificationClose.addEventListener('click', () => {
-    sessionNotification.classList.add('is-hidden')
-  })
-
-  function show (el: HTMLElement) { el.classList.remove('is-hidden') }
-  function hide (el: HTMLElement) { el.classList.add('is-hidden') }
-
-  function refreshButtons (user: User | null, gist: GistMetadata | null) {
-    if (user === null) {
-      show(signInItem); hide(accountItem)
-      hide(saveAsPublicGistControl); hide(saveAsSecretGistControl)
-      hide(updateGistControl); hide(saveAsNewGistControl)
-      return
-    }
-    hide(signInItem); show(accountItem)
-    accountAvatar.src = user.avatar_url
-    accountLogin.textContent = '@' + user.login
-
-    const ownsLoadedGist = gist !== null && gist.ownerLogin === user.login
-    if (ownsLoadedGist) {
-      hide(saveAsPublicGistControl); hide(saveAsSecretGistControl)
-      show(updateGistControl); show(saveAsNewGistControl)
-    } else {
-      show(saveAsPublicGistControl); show(saveAsSecretGistControl)
-      hide(updateGistControl); hide(saveAsNewGistControl)
-    }
-  }
-
-  currentUser$.subscribe(user => refreshButtons(user, loadedGist$.value))
-  loadedGist$.subscribe(gist => refreshButtons(currentUser$.value, gist))
-
-  signInItem.addEventListener('click', () => {
-    signIn(() => { refreshMe() })
-  })
-  signOutItem.addEventListener('click', e => {
-    e.preventDefault()
-    signOut()
-  })
-
-  function hasOutput (): boolean {
-    const out = (document.getElementById('output') as HTMLElement).textContent
-    return out !== null && out.trim().length > 0
-  }
-
-  function openSaveModal (mode: 'public' | 'secret' | 'new') {
-    pendingVisibility = mode === 'public' ? true : mode === 'secret' ? false : loadedGist$.value?.public ?? true
-    saveGistModalTitle.textContent = mode === 'public'
-      ? 'Save as public gist'
-      : mode === 'secret' ? 'Save as secret gist' : 'Save as new gist'
-    saveGistName.value = ''
-    saveGistIncludeOutput.checked = false
-    hide(saveGistNameError); hide(saveGistError); saveGistError.textContent = ''
-    if (hasOutput()) show(saveGistIncludeOutputField); else hide(saveGistIncludeOutputField)
-    saveGistModal.classList.add('is-active')
-    saveGistName.focus()
-  }
-  function closeSaveModal () {
-    saveGistModal.classList.remove('is-active')
-    pendingVisibility = null
-  }
-
-  saveAsPublicGistBtn.addEventListener('click', () => openSaveModal('public'))
-  saveAsSecretGistBtn.addEventListener('click', () => openSaveModal('secret'))
-  saveAsNewGistBtn.addEventListener('click', () => openSaveModal('new'))
-  saveGistModalClose.addEventListener('click', closeSaveModal)
-  saveGistCancel.addEventListener('click', closeSaveModal)
-  saveGistModal.querySelector('.modal-background').addEventListener('click', closeSaveModal)
-
-  function readOutputForSave (): string | undefined {
-    if (!saveGistIncludeOutput.checked) return undefined
-    return (document.getElementById('output') as HTMLElement).textContent ?? ''
-  }
-
-  function getEditorCode (): string {
-    return (document.querySelector('#code .cm-content') as any).cmView.view.state.doc.toString()
-  }
-
-  saveGistConfirm.addEventListener('click', () => {
-    const name = saveGistName.value.trim()
-    if (name === '') { show(saveGistNameError); return }
-    hide(saveGistNameError); hide(saveGistError)
-    const payload = {
-      name,
-      public: pendingVisibility ?? true,
-      code: getEditorCode(),
-      output: readOutputForSave()
-    }
-    createGist(payload).subscribe({
-      next: saved => {
-        const params = new URLSearchParams(location.search)
-        params.set('gist', saved.id)
-        const newUrl = `${location.pathname}?${params.toString()}${location.hash}`
-        history.replaceState(null, '', newUrl)
-        closeSaveModal()
-      },
-      error: err => {
-        saveGistError.textContent = err?.message ?? 'Could not save gist.'
-        show(saveGistError)
-        if (typeof err?.message === 'string' && err.message.includes('401')) {
-          notifySessionExpired()
-        }
-      }
-    })
-  })
-
-  updateGistBtn.addEventListener('click', () => {
-    const gist = loadedGist$.value
-    if (gist === null) return
-    const payload = {
-      filename: gist.filename,
-      code: getEditorCode(),
-      output: hasOutput() ? ((document.getElementById('output') as HTMLElement).textContent ?? '') : undefined
-    }
-    updateGist(gist.id, payload).subscribe({
-      next: () => { /* no-op success */ },
-      error: err => {
-        if (typeof err?.message === 'string' && err.message.includes('401')) {
-          notifySessionExpired()
-        }
-      }
-    })
-  })
-
-  // Initial check
-  refreshMe()
-}
-
 export function initView () {
   scriptExecution(executeButton, () => groovyConsole.executeScript(version.value, codeCM.getCode()))
   scriptExecution(inspectAstButton, () => groovyConsole.inspectAst(version.value, codeCM.getCode(), astPhaseSelect.value))
@@ -467,5 +302,5 @@ export function initView () {
 
   setupNavbarBurgerClickHandlers()
   setupModeSwitchersAndRestoreSavedColorMode()
-  setupAuthAndGistUi()
+  setupGistUi(codeCM)
 }
